Use AMD module format as supported by almond
It's common to use almond instead of RequireJS in production if you use the RequireJS optimizer. The benefit is a smaller download and fewer requests.

I updated the AMD module definition to use the one that is supported by almond. It is also the same format that jQuery and Backbone use.

More info available here: https://github.com/jrburke/almond#common-errors